requirements.txt
streamlit
spacy
requests
langchain
openai
python-dotenv
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
langchain_openai
